add chainID to BlockChainReactor, add chainID as GogPlugin's init parameter.

+++ x/gov/proposals.go
@@ -456,8 +456,8 @@ func (pp *ParamProposal) Verify(submitBlock uint64, blockHash common.Hash, state
 
 	var voteDuration = xcom.ParamProposalVote_DurationSeconds()
 	if chainID == big.NewInt(101) {
-		if GetCurrentActiveVersion(state) >= 2560 { // version 0.10.0
-			voteDuration = 24 * 3600 //24 hours
+		if GetCurrentActiveVersion(state) >= uint32(2560) { // version 0.10.0
+			voteDuration = uint64(24 * 3600) //24 hours
 		}
 	}
 	endVotingBlock := xutil.EstimateEndVotingBlockForParaProposal(submitBlock, voteDuration)
